Enhance AeroScoop chance display and rolling precision
The JEI display now uses a dedicated formatter for drop chances, ensuring
accurate and clean representation, especially for very small percentages,
by dynamically adjusting precision and trimming unnecessary zeros.

Additionally, the internal chance rolling mechanism has been updated to use
`nextDouble` for higher fidelity random number generation, aligning with
the improved display precision.

+++ src/main/java/dev/ftb/mods/ftbskies2aerocompanion/aeroscoop/recipe/AeroScoopResult.java
@@ -28,7 +28,7 @@ public record AeroScoopResult(Item item, int count, double chance) {
     public ItemStack roll(RandomSource rng) {
         int guaranteed = (int) Math.floor(chance);
         double bonus = chance - guaranteed;
-        int multiplier = guaranteed + (rng.nextFloat() < bonus ? 1 : 0);
+        int multiplier = guaranteed + (rng.nextDouble() < bonus ? 1 : 0);
         if (multiplier <= 0) {
             return ItemStack.EMPTY;
         }
